pyproject.toml
[tool.poetry]
name = "test_api"
version = "0.1.0"
description = ""
authors = ["adhammhaydat <adhammohidat123@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytz = "^2021.3"
djangorestframework-simplejwt = "^5.0.0"
Pillow = "^8.4.0"
django-cors-headers = "^3.10.1"
whitenoise = "^5.3.0"

[tool.poetry.dev-dependencies]
Django = "^4.0"
djangorestframework = "^3.12.4"
flake8 = "^4.0.1"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
